backend/src/requirements.txt
reboot-resemble>=0.5.3
pytest>=7.4.2
types-protobuf>=4.24.0.20240129
twilio>=7.0.0